pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "konnyaku_gpt"
version = "0.0.0"
authors = [
  { name="Trent Fellbootman", email="trentfellbootman@gmail.com" },
]
description = "AI-powered multimodal subtitle generator."
readme = "README.md"
requires-python = ">=3.11,<4.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
  'srt>=3.5.3,<4.0.0',
  'moviepy>=1.0.3,<2.0.0',
  'imageio>=2.31.1,<3.0.0',
  'imageio-ffmpeg>=0.4.8,<0.5.0',
  'proglog>=0.1.10,<0.2.0',
  'numpy>=1.24.3,<2.0.0',
  'requests>=2.31.0,<3.0.0',
  'librosa>=0.10.1,<0.11.0',
  'pillow>=9.4.0,<10.0.0',

  'torch>=2.0.0,<3.0.0',
  'transformers>=4.32.1,<5.0.0',
  'openai>=0.27.4,<0.28.0',
]

[project.urls]
"GitHub" = "https://github.com/Trent-Fellbootman/konnyaku-gpt"